fix: align urlscan sdk actions with native outputs
